Refactor and bug fix
